Bit of a refactor to help with future cutover removing dependencies between updateTemplate new/old
